[Feature #157933384] Update directory structure
Remove collections from __init__.py and remove config classes from
config/__init__.py
